LibUIDK: List控件绘制垂直网格线时，支持从第一行开始绘制。List控件支持奇偶行不同背景色；X64支持.

+++ UIShop/PropertiesWnd.cpp
@@ -40,7 +40,7 @@ int InitWindowIDOptions(CMFCPropertyGridProperty *pProp, BOOL bIncludeSelf)
 	CString strCurWinID = (LPCTSTR)pWndProp->m_strID;
 
 	std::vector<UIWNDPROPERTIES *> *pWindows = CUIMgr::GetUIWndPropertiesList();
-	int nCount = pWindows->size();
+	size_t nCount = pWindows->size();
 	for (int i = 0; i < nCount; ++i)
 	{
 		UIWNDPROPERTIES *pWindow = (*pWindows)[i];
